Containerize cerberus
This commit adds support to run the tool as a container on the host
with access to kubeconfig for better portability. The plan is to add
a webhook to trigger image builds on every commit to this repo at
quay.io to make sure it has the latest code instead of the user
taking care of building the image manually.