Added support for cerberus cop
This commit enables cop id's to be assigned for each day in config
file. The cop for the day is tagged while reporting cluster component
failures. If a cop isn't assigned or if the assinged cop isn't a
member of the slack channel or if a wrong slack member id is
provided, everyone on the channel is notified by using @here tag.