Added support to report failed containers
This commits enables the status of the containers in the pods to
be checked in addition to the pod status to determine go/no-go.
Containers failing in a pod in addition to the pod name are
logged in the report, thereby enabling the user to debug the
failed pods easily.