Time the watch operations for each iteration
This commit adds support to time the operations like watching nodes,
cluster operators, namespaces e.t.c and prints out the info after
each iteration. This will help in understanding how long each operation
is taking and act accordingly.

It will be helpful in looking at how long cerberus takes to run the
checks as wells as the impact on the timing when running it on a small
scale vs large scale clusters. Eventually they can be exposed as metrics
and act accordingly if higher latency is observed with Kube/OpenShift API.