Alert on high Kube API server latencies
This commit will enable Cerberus to query prometheus if deployed
for KubeAPILatencyHigh alert at the end of each iteration and warns
the user if 99th percentile latency for given requests to the
kube-apiserver is above 1 second. It's the official SLI/SLO defined
for Kubernetes. Higher latencies is an indicator of problem in the
cluster.

Fixes https://github.com/openshift-scale/cerberus/issues/65.