Reduce the load on API server for bulk requests
This commit enables Cerberus to handle bulk requests in chunks
instead of a single go as it can generate a very large response
when there are thousands of objects running in the cluster and
overload the server by consuming a large amount of resources.
The chunk size is set to 250 by default to break the large request
while still preserving the consistency of the total request and
improve responsiveness.